# app/view/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Atendimento API",
    "description": "Swagger para a API de Atendimento do VetEmCasa.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "karin.ingwersen96@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/listar_atendimentos": {
      "get": {
        "description": "Retorna todos os atendimentos.",
        "operationId": "findAtendimentos",
        "produces": [
          "application/json",
          "application/xml",
          "text/xml",
          "text/html"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "resposta do atendimento",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Atendimento"
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/v1/detalhes_atendimento": {
      "get": {
        "description": "Retorna os detalhes de um atendimento de acordo com o seu ID",
        "operationId": "detailsAtendimento",
        "produces": [
          "application/json",
          "application/xml",
          "text/xml",
          "text/html"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do atendimento para buscar",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "pet response",
            "schema": {
              "$ref": "#/definitions/Atendimento"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    },
    "/v1/criar_atendimento": {
      "post": {
          "description": "Cria um novo atendimento",
          "operationId": "addAtendimento",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "atendimento",
              "in": "body",
              "description": "Atendimento a ser criado.",
              "required": true,
              "schema": {
                "$ref": "#/definitions/NewAtendimento"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "atendimento criado"
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/ErrorModel"
              }
            }
          }
      }
    },
    "/v1/apagar_atendimento": {
      "delete": {
          "description": "Deleta um atendimento",
          "operationId": "deleteAtendimento",
          "produces": [
            "application/json"
          ],
           "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do atendimento para apagar",
            "required": true,
            "type": "string"
          }
        ],
          "responses": {
            "204": {
              "description": "atendimento deletado"
            },
            "default": {
              "description": "unexpected error",
              "schema": {
                "$ref": "#/definitions/ErrorModel"
              }
            }
          }
      }
    },
    "/v1/atualizar_atendimento": {
      "put": {
        "description": "Atualiza um atendimento",
        "operationId": "updateAtendimento",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do atendimento para atualizar",
            "required": true,
            "type": "string"
          },
          {
            "name": "atendimento",
            "in": "body",
            "description": "Atendimento a ser atualizado.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewAtendimento"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "atendimento atualizado"
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Atendimento": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/NewAtendimento"
        },
        {
          "required": [
            "id"
          ],
          "properties": {
            "id": {
              "type": "string"
            }
          }
        }
      ]
    },
    "NewAtendimento": {
      "type": "object",
      "required": [
        "pet",
        "tutor",
        "veterinario",
        "data",
        "horario"
      ],
      "properties": {
        "pet": {
          "type": "string"
        },
        "tutor": {
          "type": "string"
        },
        "veterinario": {
          "type": "string"
        },
        "data": {
          "type": "string"
        },
        "horario": {
          "type": "string"
        }
      }
    },
    "ErrorModel": {
      "type": "object",
      "required": [
        "code",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}